+++ src/Hasklepias/Reexports.hs
@@ -13,10 +13,10 @@ module Hasklepias.Reexports (
 
     -- * Re-exports
       module GHC.Num
-    , module GHC.Generic
+    , module GHC.Generics
+    , module GHC.Show
     , module Control.Monad
     , module Control.Applicative
-    , module Data.Aeson
     , module Data.Bool
     , module Data.Either
     , module Data.Eq
@@ -26,7 +26,6 @@ module Hasklepias.Reexports (
     , module Data.Functor
     , module Data.List
     , module Data.Ord
-    , module Data.Show
     , module Data.Time.Calendar 
     , module Data.Text
     , module Data.Tuple
@@ -36,10 +35,10 @@ module Hasklepias.Reexports (
 ) where
 
 import safe GHC.Num                         ( Integer )
-import safe GHC.Generic                     ( Generic )
+import safe GHC.Generics                    ( Generic )
+import safe GHC.Show                        ( Show(..) )
 import safe Control.Monad                   ( (=<<), Functor(fmap) )
 import safe Control.Applicative             ( (<$>) )
-import Data.Aeson                           ( ToJSON(..) )
 import safe Data.Bool                       ( Bool(..)
                                             , (&&), not, (||)
                                             , bool
@@ -68,7 +67,6 @@ import safe Data.Maybe                      ( Maybe(..),
                                               maybeToList )
 import safe Data.Ord                        ( Ord((>=), (<), (>), (<=))
                                             , max, min )
-import safe Data.Show                       ( Show(..) )
 import safe Data.Time.Calendar              ( Day, MonthOfYear, Year
                                             , fromGregorian
                                             , diffDays )